Struct google_container1_beta1::MasterAuth

pub struct MasterAuth {
    pub bearer_token: Option<String>,
    pub password: Option<String>,
    pub client_certificate: Option<String>,
    pub user: Option<String>,
    pub client_key: Option<String>,
    pub cluster_ca_certificate: Option<String>,
}

The authentication information for accessing the master. Authentication is either done using HTTP basic authentication or using a bearer token.

This type is not used in any activity, and only used as part of another schema.

Fields

bearer_token

The token used to authenticate API requests to the master. The token is to be included in an HTTP Authorization Header in all requests to the master endpoint. The format of the header is: "Authorization: Bearer ".

password

The password to use for HTTP basic authentication when accessing the Kubernetes master endpoint. Because the master endpoint is open to the internet, you should create a strong password.

client_certificate

[Output only] Base64 encoded public certificate used by clients to authenticate to the cluster endpoint.

user

The username to use for HTTP basic authentication when accessing the Kubernetes master endpoint.

client_key

[Output only] Base64 encoded private key used by clients to authenticate to the cluster endpoint.

cluster_ca_certificate

[Output only] Base64 encoded public certificate that is the root of trust for the cluster.