Struct google-container1_beta1::Operation

pub struct Operation {
    pub status: String,
    pub name: String,
    pub zone: String,
    pub error_message: String,
    pub target_link: String,
    pub operation_type: String,
    pub self_link: String,
    pub target: String,
}

Defines the operation resource. All fields are output only.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

zones operations get projects (response)

zones clusters delete projects (response)

zones clusters create projects (response)

Fields

status

The current status of the operation.

name

The server-assigned ID for this operation. If the operation is fulfilled upfront, it may not have a resource name.

zone

The name of the Google Compute Engine zone in which the operation is taking place.

error_message

If an error has occurred, a textual description of the error.

target_link

Server-defined URL for the target of the operation.

operation_type

The operation type.

self_link

Server-defined URL for the resource.

target

[Optional] The URL of the cluster resource that this operation is associated with.