Struct google_container1_beta1::ProjectMethods

pub struct ProjectMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on project resources. It is not used directly, but through the Container hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_container1_beta1 as container1_beta1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use container1_beta1::Container;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Container::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `clusters_list(...)`, `operations_list(...)`, `zones_clusters_create(...)`, `zones_clusters_delete(...)`, `zones_clusters_get(...)`, `zones_clusters_list(...)`, `zones_operations_get(...)` and `zones_operations_list(...)`
// to build up your call.
let rb = hub.projects();

Methods

impl<'a, C, A> ProjectMethods<'a, C, A>

fn zones_clusters_get(&self, project_id: &str, zone_id: &str, cluster_id: &str) -> ProjectZoneClusterGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a specific cluster.

Arguments

projectId - The Google Developers Console project ID or project number.

zoneId - The name of the Google Compute Engine zone in which the cluster resides.

clusterId - The name of the cluster to retrieve.

fn operations_list(&self, project_id: &str) -> ProjectOperationListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all operations in a project, across all zones.

Arguments

projectId - The Google Developers Console project ID or project number.

fn zones_clusters_delete(&self, project_id: &str, zone_id: &str, cluster_id: &str) -> ProjectZoneClusterDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the cluster, including the Kubernetes master and all worker nodes.

Firewalls and routes that were configured at cluster creation are also deleted.

Arguments

projectId - The Google Developers Console project ID or project number.

zoneId - The name of the Google Compute Engine zone in which the cluster resides.

clusterId - The name of the cluster to delete.

fn clusters_list(&self, project_id: &str) -> ProjectClusterListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all clusters owned by a project across all zones.

Arguments

projectId - The Google Developers Console project ID or project number.

fn zones_operations_get(&self, project_id: &str, zone_id: &str, operation_id: &str) -> ProjectZoneOperationGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the specified operation.

Arguments

projectId - The Google Developers Console project ID or project number.

zoneId - The name of the Google Compute Engine zone in which the operation resides. This is always the same zone as the cluster with which the operation is associated.

operationId - The server-assigned name of the operation.

fn zones_operations_list(&self, project_id: &str, zone_id: &str) -> ProjectZoneOperationListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all operations in a project in a specific zone.

Arguments

projectId - The Google Developers Console project ID or project number.

zoneId - The name of the Google Compute Engine zone to return operations for.

fn zones_clusters_list(&self, project_id: &str, zone_id: &str) -> ProjectZoneClusterListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all clusters owned by a project in the specified zone.

Arguments

projectId - The Google Developers Console project ID or project number.

zoneId - The name of the Google Compute Engine zone in which the cluster resides.

fn zones_clusters_create(&self, request: CreateClusterRequest, project_id: &str, zone_id: &str) -> ProjectZoneClusterCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a cluster, consisting of the specified number and type of Google Compute Engine instances, plus a Kubernetes master instance.

The cluster is created in the project's default network.

A firewall is added that allows traffic into port 443 on the master, which enables HTTPS. A firewall and a route is added for each node to allow the containers on that node to communicate with all other instances in the cluster.

Finally, an entry is added to the project's global metadata indicating which CIDR range is being used by the cluster.

Arguments

request - No description provided.

projectId - The Google Developers Console project ID or project number.

zoneId - The name of the Google Compute Engine zone in which the cluster resides.