Struct google-container1_beta1::Cluster

pub struct Cluster {
    pub status: String,
    pub container_ipv4_cidr: String,
    pub description: String,
    pub node_routing_prefix_size: i32,
    pub master_auth: MasterAuth,
    pub cluster_api_version: String,
    pub creation_timestamp: String,
    pub name: String,
    pub endpoint: String,
    pub network: String,
    pub zone: String,
    pub num_nodes: i32,
    pub node_config: NodeConfig,
    pub self_link: String,
    pub status_message: String,
    pub services_ipv4_cidr: String,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

zones clusters get projects (response)

Fields

status

[Output only] The current status of this cluster.

container_ipv4_cidr

[Output only] The IP addresses of the container pods in this cluster, in CIDR notation (e.g. 1.2.3.4/29).

description

An optional description of this cluster.

node_routing_prefix_size

[Output only] The size of the address space on each node for hosting containers.

master_auth

The HTTP basic authentication information for accessing the master. Because the master endpoint is open to the internet, you should create a strong password.

cluster_api_version

The API version of the Kubernetes master and kubelets running in this cluster. Leave blank to pick up the latest stable release, or specify a version of the form "x.y.z". The Google Container Engine release notes lists the currently supported versions. If an incorrect version is specified, the server returns an error listing the currently supported versions.

creation_timestamp

[Output only] The time the cluster was created, in RFC3339 text format.

name

The name of this cluster. The name must be unique within this project and zone, and can be up to 40 characters with the following restrictions: - Lowercase letters, numbers, and hyphens only. - Must start with a letter. - Must end with a number or a letter.

endpoint

[Output only] The IP address of this cluster's Kubernetes master. The endpoint can be accessed from the internet at https://username:password@endpoint/.

See the masterAuth property of this resource for username and password information.

network

The name of the Google Compute Engine network to which the cluster is connected.

zone

[Output only] The name of the Google Compute Engine zone in which the cluster resides.

num_nodes

The number of nodes to create in this cluster. You must ensure that your Compute Engine resource quota is sufficient for this number of instances plus one (to include the master). You must also have available firewall and routes quota.

node_config

The machine type and image to use for all nodes in this cluster. See the descriptions of the child properties of nodeConfig.

self_link

[Output only] Server-defined URL for the resource.

status_message

[Output only] Additional information about the current status of this cluster, if available.

services_ipv4_cidr

[Output only] The IP addresses of the Kubernetes services in this cluster, in CIDR notation (e.g. 1.2.3.4/29). Service addresses are always in the 10.0.0.0/16 range.